+++ extensions/git/src/components/views/BranchesView.tsx
@@ -109,10 +109,9 @@ export function BranchesView(context: RepositoryContext & NavigationContext) {
 }
 
 function BranchListItem(context: RepositoryContext & NavigationContext & { branch: Branch }) {
-  const hasConflicts =
-    context.branch.type === "current" && context.status.data?.files?.some((file) => file.isConflicted);
+  const hasConflicts = context.branch.type === "current" && context.status.data.files.some((file) => file.isConflicted);
 
-  const hasUncommittedChanges = context.branch.type === "current" && context.status.data?.files?.length !== 0;
+  const hasUncommittedChanges = context.branch.type === "current" && context.status.data.files.length !== 0;
 
   const accessories: List.Item.Accessory[] = useMemo(() => {
     const result = [];
@@ -308,8 +307,8 @@ function BranchListItem(context: RepositoryContext & NavigationContext & { branc
 }
 
 function DetachedHeadListItem(context: RepositoryContext & NavigationContext & { detachedHead: DetachedHead }) {
-  const hasConflicts = context.status.data?.files?.some((file) => file.isConflicted);
-  const hasUncommittedChanges = context.status.data?.files?.length !== 0;
+  const hasConflicts = context.status.data.files.some((file) => file.isConflicted);
+  const hasUncommittedChanges = context.status.data.files.length !== 0;
 
   const accessories: List.Item.Accessory[] = useMemo(() => {
     const result: List.Item.Accessory[] = [];